Add project for the AD9136-FMC-EBZ + ZC706
Add a project for the AD9136-FMC-EBZ dual DAC board connected to the ZC706
carrier board.

This project assumes that the transceiver reference clock and SYSREF are
provided via the clock distribution chip that is found on the
AD9136-FMC-EBZ board. The reference clock rate must be the lane rate
divided by 40. SYSREF should be the reference clock divided by 64 (or a
larger power-of-two).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>